Move translation code into V2ToFHIR Module
